Quiet compiler warnings
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@4351 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/util/xp/ut_stringbuf.cpp
@@ -172,7 +172,7 @@ void UT_Stringbuf::grow_nocopy(size_t n)
 	++n;	// allow for zero termination
 	if (n > capacity()) {
 		const size_t nCurSize = size();
-		n = priv_max(n, nCurSize * g_rGrowBy);
+		n = priv_max(n, (size_t)(nCurSize * g_rGrowBy));
 		delete[] m_psz;
 		m_psz  = new char_type[n];
 		m_pEnd = m_psz + nCurSize;
@@ -185,7 +185,7 @@ void UT_Stringbuf::grow_copy(size_t n)
 	++n;	// allow for zero termination
 	if (n > capacity()) {
 		const size_t nCurSize = size();
-		n = priv_max(n, nCurSize * g_rGrowBy);
+		n = priv_max(n, (size_t)(nCurSize * g_rGrowBy));
 		char_type* p = new char_type[n];
 		if (m_psz) {
 			memcpy(p, m_psz, size() + 1);
